bot/requirements.txt
web3
python-dotenv
eth-account